Original MS-DOS Acidwarp 4.10 by Noah Spurrier and Mark Bilk
These are the contents of http://www.noah.org/acidwarp/acidwarp.zip
with only changes being lowercased filenames and Unix line endings.
Line endings in vc++70.mak were not converted because that file is
for DOS only.

The original unaltered acidwarp.exe binary is included. It was
built with an old proprietary compiler, and this source may not
build with any modern toolchain. The binary can run in DOSBox or
other emulators.