X11 XPCE port by Boris Gjenero
This was used to develop libXPCE, an X11 PseudoColor emulation
library. It was used for running 256 colour programs when
the PseudoColour visual class is unavailable. That old library
is obsolete because SDL is better. It has not been included,
and so xacidwarp cannot be built from this source. This version
only exists to document the transition from SVGALib to SDL.