Disable Windows full screen mode palette hack.
Windows normally doesn't allow an application to access all 256
colours of the palette in 256 colour mode. There was a workaround,
but it was messy and it didn't work anymore in Windows 7. In the
future, a more reliable solution would be not using 256 colour
mode and emulating colour cycling by repeated blitting.