Rewrite SDL for output at native bit depth.
This fixes the Windows problems by not using 8 bpp for full screen
mode as long as the desktop is at a higher bit depth. There is also
various cleanup and renaming of SDL code.

In typical desktop situations, images will be drawn in an 8 bpp
SDL_Surface, which will then be copied via SDL_BlitSurface() to
a native depth screen. If scaling is enabled, drawing is instead
to buf_graf, which is first scaled up to the surface. With a
palette, drawing is to buf_graf, which is then scaled or copied
directly to the screen.

This code now responds to expose events, which may be important
when using a palette in some environments, because the image is
then unchanging while the palette cycles.