Add option for scaling images with floating point image generator.
This adds an option for scaling images so that they look like
they would look at the original VGA resolution of 320x200. Some
images are unaffected.

Many already divided by xmax and ymax. Those were already scaled,
and they continue to scale the same way. Those that directly
access buf_graf are not scaled.

Scaling makes the right seam visible on some images when a
sparkly pallette is used. This still needs to be fixed.
There is a problem when drawing image 10 using floating point.