Add command line options for full screen mode and change options.
This creates new command line options for full screen mode and
for using desktop resolution in full screen mode. The floating
point image generator becomes the default. The -f option is used
for full screen mode because that is obvious, and the new -o
option can be used to use the classic image generator.