Support full screen with SDL 2 in a simplified way
Instead of scanning the list of modes and finding an appropriate
one, it may be enough to simply use SDL_SetWindowFullscreen() and
let it find an appropriate mode. This doesn't take advantage of
the scaling that display.c is able to do.

Tested in Windows 7. Doesn't work for me in Linux because SDL 2
full screen is broken in general.