Fix colour cycling timer code, which was wrong but worked
Colour cycling code was embarrassingly wrong, not unlocking the
mutex. The SDL 1.2 documentation isn't exactly clear on this, just
saying that it the mutex needs to be locked before the cond wait.
However, SDL 2 documentation is clear, and SDL 1.2 directly uses
phtreads, so it is clear that the mutex will be re-locked after
the wait, and needs to be unlocked. The resulting code seems far
too big for what it does.